+++ keras_cv/bounding_box/utils.py
@@ -92,6 +92,26 @@ def clip_to_image(bounding_boxes, images, bounding_box_format):
     return clipped_bounding_boxes
 
 
+# TODO (tanzhenyu): merge with clip_to_image
+def _clip_boxes(boxes, box_format, image_shape):
+    """Clip boxes to the boundaries of the image shape"""
+    if boxes.shape[-1] != 4:
+        raise ValueError(
+            "boxes.shape[-1] is {:d}, but must be 4.".format(boxes.shape[-1])
+        )
+
+    if isinstance(image_shape, list) or isinstance(image_shape, tuple):
+        height, width, _ = image_shape
+        max_length = [height, width, height, width]
+    else:
+        image_shape = tf.cast(image_shape, dtype=boxes.dtype)
+        height, width, _ = tf.unstack(image_shape, axis=-1)
+        max_length = tf.stack([height, width, height, width], axis=-1)
+
+    clipped_boxes = tf.math.maximum(tf.math.minimum(boxes, max_length), 0.0)
+    return clipped_boxes
+
+
 def _format_inputs(boxes, images):
     boxes_rank = len(boxes.shape)
     if boxes_rank > 3:
