+++ keras_cv/models/object_detection/faster_rcnn.py
@@ -15,8 +15,8 @@
 import tensorflow as tf
 
 from keras_cv import bounding_box
-from keras_cv.bounding_box.converters import _clip_boxes
 from keras_cv.bounding_box.converters import _decode_deltas_to_boxes
+from keras_cv.bounding_box.utils import _clip_boxes
 from keras_cv.layers.object_detection.anchor_generator import AnchorGenerator
 from keras_cv.layers.object_detection.roi_align import _ROIAligner
 from keras_cv.layers.object_detection.roi_generator import ROIGenerator
